mudanças para remover warnings

+++ going/databases/show_event.php
@@ -45,35 +45,31 @@ include_once("templates/show_event.php");
 <br>
 
 <div class="nav_buttons">
-        <?php if ($logged) { ?>
-            <?php if (!isRegisteredInEvent($id, $_SESSION['username'])) {?>
+    <?php if ($logged) { ?>
+        <?php if (!isRegisteredInEvent($id, $_SESSION['username'])) {?>
 
-			<button type="button" onclick="location.href='action_register_event.php?id=<?=$id?>'">Register</button>
+            <button type="button" onclick="location.href='action_register_event.php?id=<?=$id?>'">Register</button>
 
-            <?php } else {?>
-			<button type="button" onclick="location.href='action_unregister_event.php?id=<?=$id?>'">Unregister</button>
-
-            <?php } ?>
-
-            <?php if ($_SESSION['username'] == $event['creator']
-                                           || $_SESSION['username'] == 'admin') { ?>
-			<button type="button" onclick="location.href='action_delete_event.php?id=<?=$id?>'">Delete</button>
-
-            <?php } ?>
-
-			 <?php if ($_SESSION['username'] == $event['creator']
-                                           || $_SESSION['username'] == 'admin') { ?>
-			<button type="button" onclick="location.href='read_edit_event.php?id=<?=$id?>'">Edit</button>
-
-            <?php } ?>
-
-
+        <?php } else {?>
+            <button type="button" onclick="location.href='action_unregister_event.php?id=<?=$id?>'">Unregister</button>
 
         <?php } ?>
-<button type="button" onclick="location.href='list_events.php'">Back</button>
+
+        <?php if ($_SESSION['username'] == $event['creator']
+            || $_SESSION['username'] == 'admin') { ?>
+            <button type="button" onclick="location.href='action_delete_event.php?id=<?=$id?>'">Delete</button>
+
+        <?php } ?>
+
+        <?php if ($_SESSION['username'] == $event['creator']
+            || $_SESSION['username'] == 'admin') { ?>
+            <button type="button" onclick="location.href='read_edit_event.php?id=<?=$id?>'">Edit</button>
+
+        <?php } ?>
+
+
+
+    <?php } ?>
+    <button type="button" onclick="location.href='list_events.php'">Back</button>
 </div>
-</div>
-</header>
-</body>
-</html>
